Ai_image-interpretator/requirements.txt
google-genai
python-dotenv
SpeechRecognition
gTTS
pyaudio
opencv-python
